+++ minillm/transformers/src/transformers/models/qwen_parallel/modeling_qwen_paralle.py
@@ -562,9 +562,15 @@ class ParallelQWenAttention(nn.Module):
                         attention_mask = attention_mask.masked_fill(~causal_mask, torch.finfo(query.dtype).min)
                 else:
                     attention_mask = causal_mask
+                # attn_output = F.scaled_dot_product_attention(
+                #     query, key, value, attn_mask=attention_mask
+                # ).transpose(1, 2)
                 attn_output = F.scaled_dot_product_attention(
-                    query, key, value, attn_mask=attention_mask
-                ).transpose(1, 2)
+                    query.to(torch.float32),
+                    key.to(torch.float32),
+                    value.to(torch.float32),
+                    attn_mask=attention_mask.to(torch.float32)
+                ).to(query.dtype).transpose(1, 2)
                 attn_weight = None
             else:
                 attn_output, attn_weight = self._attn(
